+++ CHANGELOG.md
@@ -9,8 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `Beaker.dataset.size()`.
+- Added `Beaker.dataset.rename()`.
+
+## [v0.5.3](https://github.com/allenai/beaker-py/releases/tag/v0.5.3) - 2022-04-08
+
+### Added
+
 - Added `Beaker.dataset.commit()`.
 - Added `Beaker.dataset.ls()`.
+- Added `Beaker.dataset.stream_file()`.
 
 ### Changed